lambda/api/requirements.txt
fastapi==0.109.0
mangum==0.17.0
boto3==1.34.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.23.3

